Split rcl_bindings.cpp file
