+++ binding.gyp
@@ -23,8 +23,20 @@
         './src/handle_manager.cpp',
         './src/rcl_action_bindings.cpp',
         './src/rcl_bindings.cpp',
+        './src/rcl_client_bindings.cpp',
+        './src/rcl_context_bindings.cpp',
+        './src/rcl_graph_binding.cpp',
+        './src/rcl_guard_condition_bindings.cpp',
         './src/rcl_handle.cpp',
         './src/rcl_lifecycle_bindings.cpp',
+        './src/rcl_logging_bindings.cpp',
+        './src/rcl_names_bindings.cpp',
+        './src/rcl_node_bindings.cpp',
+        './src/rcl_publisher_bindings.cpp',
+        './src/rcl_service_bindings.cpp',
+        './src/rcl_subscription_bindings.cpp',
+        './src/rcl_time_point_bindings.cpp',
+        './src/rcl_timer_bindings.cpp',
         './src/rcl_utilities.cpp',
         './src/shadow_node.cpp',
       ],
